Improve performance by defering draw after all events in current frame
If you spam many keys quickly, what could happen is that they are all buffered
and have to catch up by generating a frame for each keypress. This is because
displaying a frame will cause it to block for vsync on many devices, and
generating a frame might be a hard task in general. This commit will improve
the performance by defering the main_draw until after all events received
while the thread was potentially blocked are processed.